+++ src/cli/index.ts
@@ -4,29 +4,32 @@ import getStricter from '../factory';
 export default (): number => {
     const argv = yargs
         .version(process.env.STRICTER_VERSION as string)
-        .option('c', {
-            alias: 'config',
+        .option('config', {
+            alias: 'c',
             description: 'Specify config location',
             string: true,
             requiresArg: true,
         })
-        .option('r', {
-            alias: 'reporter',
+        .option('reporter', {
+            alias: 'r',
             description: 'Specify reporter',
             choices: ['console', 'mocha', 'junit'],
             requiresArg: true,
         })
-        .option('v', {
-            alias: 'verify',
+        .option('rule', {
             description: 'Verify particular rule',
             array: true,
             requiresArg: true,
+        })
+        .option('clearCache', {
+            description: 'Clears cache',
         }).argv;
 
     const stricter = getStricter({
         config: argv.config,
         reporter: argv.reporter,
-        rulesToVerify: argv.verify,
+        rulesToVerify: argv.rule,
+        clearCache: argv.clearCache,
     });
 
     const result = stricter();