+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Use `rollup` instead of `fusebox` for bundling
 - Fix default rules handling
 - Refactoring
+- Cache dependency parsing
+- Add command to clear cache (`--clearCache`)
 
 
 ## 0.2.0 - More correct